requirements.txt
Cython==3.0.11
numpy==1.26.4
scipy
torch>=2.2.2
torchvision
torchaudio
scikit-learn
wheel
setuptools
aeon
numba==0.61.2
cupy-cuda12x==13.4.1
psutil
tqdm==4.67.1
jax[cuda12]
